Revert one CodingStyle change
Without this rule strictly enforced, it quickly becomes harder and
harder to remember where each class is and for an outsider to quickly
get to some particular peace of code.

The rest of Marc-Andre's preferences I can learn to accept and maybe even
appreciate in time but this is one of the over-my-dead-body thing.