Only apply post-install config when installation is done
For domains we create ourselves, we now check for signs of OS
installation after (every) shutdown. If we find any, only then we apply
the post-install configuration (boot installed OS).

For pure live media, this additionally implies that its domain is
destroyed automatically on first shutdown, unless it installed an OS.

https://bugzilla.gnome.org/show_bug.cgi?id=665254