Use floppy image for only Windows express install
The underlying image remains the same for both Fedora and Windows but
now we tell libvirt to emulate a USB drive using that image for Fedora.
According to Lucas Meneghel Rodrigues <lmr@redhat.com>, floppy support
on Linux isn't reliable enough to depend on and having a floppy drive
can easily lead to weird kernel crashes.

https://bugzilla.gnome.org/show_bug.cgi?id=666956