Add devices based on what is supported by the OS
Add video, sound and tablet devices to new VM based on what is supported
by the OS in question. The information about device support comes from
libosinfo.

Note that we don't add these devices if OS itself or its support for the device
in question is unknown and rely on libvirt to fill in the safe defaults for us.

https://bugzilla.gnome.org/show_bug.cgi?id=666739