express: No need for floppy image
libvirt+qemu can simply expose a directory as a device so we don't
really need to keep/create a floppy image for each express installation.

According to Paolo Bonzini, we are safe to depend on this qemu feature
as long as we use the disk read-only and thats what we do.

https://bugzilla.gnome.org/show_bug.cgi?id=667895