Move set_app_menu to app.startup.connect_after
Boxes fails to start with a current git snapshot of gtk/glib because
set_app_menu is not called from the app startup function.
This used to work when the menu stuff was still using dbus but broke when it
switched to x properties.

Here is some non-committed documentation about this:
https://bugzilla.gnome.org/show_bug.cgi?id=668203

I'm attaching a patch that moves the menu to app startup. I've decided to put
it into the startup directly as it was not part of setup_ui() before, but it
might make sense to move it there (before creating the window).

This fixes https://bugzilla.gnome.org/show_bug.cgi?id=668584