More generic notification bar
Now its also possible to display errors through notification bar and new
types of messages are now trivial to add. It is also now visible in all
UI states except for 'DISPLAY'. Do we want it there?

Not sure if this is an issue for us but the colors of Gtk.InfoBar remain
always the same regardless of type of message. While we do set the
appropriate message type on the Gtk.InfoBar, we set our own style on it
as well. So if we need this, we need to figure how to set different
styles for different message types.

https://bugzilla.gnome.org/show_bug.cgi?id=668792