wizard: Properly handle errors on 'prepare' & 'create'
Display an error in notification and in case of 'prepare', we go back to
previous step since user is very likely to be able to fix the issue.

The error message displayed at failure from 'create' is not very useful
but we don't want to show cryptic error messages to user and we really
should aim to make it (almost) impossible to get errors from 'create'.

https://bugzilla.gnome.org/show_bug.cgi?id=668792