wizard: Present bootable ISO files on 'Source' page
Since for each addition/deletion Tracker, you get an integer ID from
Tracker against which you use a Sparql query to check what Tracker is
talking about, the code to handle the these signals will be inefficient
and unnecessarily complicated. So instead of keeping a list of medias (HW
and ISO images), we now:

1. Asynchronously fetch list of medias on startup and create the widgets
for them in the wizard.

2. When the 'Source' page of wizard is activated/shown, we refetch the
list (again asynchronously) and add/remove widgets accordingly.

https://bugzilla.gnome.org/show_bug.cgi?id=666373