First start experience improvement
Currently, when the user starts Boxes for the first time, he ends up
in an empty collection without message on what to do. Instead,
starting with the Wizard is slightly better. Proposed patch attached.

Note: there is no yet final decision from the design team on what
should be done, but only one line of the proposed patch is actually
implementing the first-time experience policy, and thus could easily
be changed if needed.

https://bugzilla.gnome.org/show_bug.cgi?id=670539