Fix a connection race and code silliness
I noticed a bug when a VM is started, it may receive two connections
very quickly, but since channel events are asynchrnous, we would
disconnect and hide the display, even if the second connection
succeeded.

The current code is silly, trying to connect automatically when needed
when the _connect_display member is set.

Get rid of that, and make sure we try to connect only once. Do not
connect/disconnect/reconnect etc..

https://bugzilla.gnome.org/show_bug.cgi?id=670539