add --checks option to troubleshot for vt/kvm
Add a simple argument --check to do some basic capability checking.
Although we are in freeze time, it can be helpful to help people who
expect boxes to work on their machine explain why it doesn't.

In a future version of boxes, we should have a translatable and
friendlier version of this warning, with a documentation explaining
how to enable VT in the bios and to load KVM.

https://bugzilla.gnome.org/show_bug.cgi?id=671673