Don't wait for domain creation event
libvirt doesn't seem to be very reliable on prolonged sessions so the
lesser we depend on it, the better. This patch is mainly a work-around
for a very hard to debug/reproduce libvirt issue: It stops sending
domain events after libvirtd has been running for a long time (where the
definition of 'long time' doesn't seem to be predictable).

OTOH, this patch doesn't do any harm and actually reduces the time it
takes before user hitting 'create' button in wizard and new machine
appearing in the collection view.

https://bugzilla.gnome.org/show_bug.cgi?id=671847