express,fedora: Install from remote repo for <= F16
Not sure if this is a good thing in general but this is the quickest
solution to this bug I could come-up with:

https://bugzilla.redhat.com/show_bug.cgi?id=804834

One issue with this change is that it makes Fedora 16 (and older) express
installation to require Internet access. On the bright side, we can now
have spice-vdagent package installed automatically from the start.

https://bugzilla.gnome.org/show_bug.cgi?id=672420