Fix dragging window in win7 shouldn't bring topbar
Win7 (and perhaps Vista), have that window management ui that allows to
maximize a window by dragging it to the top of the screen and releasing (when
there is a blue overlay over the desktop). The window will then be maximzed.

Currently, with Boxes, this is very hard to work with, Since the toolbar will
be bring up when reaching the top, and will catch the release event. Since it's
along the whole width of the screen there is no way to avoid the toolbar and
use that feature.

I think a reasonable solution would be to not bring the topbar when the button
has been pressed (another solution could be to send the release event to the
guest if it received the press event before the toolbar, but this is confusing
imho)

https://bugzilla.gnome.org/show_bug.cgi?id=674361