win7: Use drivers from extra iso at install time
If the extra driver iso is available when creating a new win7 box,
we can use these drivers during installation to be able to use
VirtIO through viostor during the installation, which is desirable
for better performance.

If the ISO is not available, this entry is silently ignored,
so it doesn't hurt to have it always there.

https://bugzilla.gnome.org/show_bug.cgi?id=679752