installer: Insert extra iso if available
This iso will be used during automatic Windows installations. It
contains the viostor driver to be able to use virtio at install time
and extra drivers (vioserial, ...) and utilities (vdagent) which are
added to the install after it's done. This improves VM integration
a lot.
The name of the iso is win-tools.iso, and it's currently looked up
in $XDG_CACHE_HOME/gnome-boxes/ first, and if not found, it's
searched in $datadir/gnome-boxes/unattended.

https://bugzilla.gnome.org/show_bug.cgi?id=679752