Turn LibvirtMachine.state from property to a method
This was getting called in multiple places as if it was a cheap method, while
in reality its a blocking rpc call. Sometimes it was even used multiple times
within a method.

The new method is called get_state_sync() to make it obvious that this is a
blocking call.

https://bugzilla.gnome.org/show_bug.cgi?id=675663