Create VM before 'review' page
This means that we now move most of the failure points before arriving to
review page. If user hits 'create' on the review page, the VM is simply
launched. OTOH if user hits 'back' or 'cancel', the created VM is deleted
along with its volume (cheap operation).

Why do we need this change?

If you look at the later commit, you'll notice that we at least need to
create the configuration(s) before 'review' since we need information
about the new domain to display to user. Now its an interesting question
whether we should only create configuration before hand or the whole
machine.

The reasons I decided to go with the latter approach was:

1. I recalled mccann saying that we should ensure that we have everything
ready at the 'review' page. I agree with that since its better to error
out during preparation and setup rather than after we told user that
everything is ready, she/he clicks create and we go *boom*.

2. IMHO it makes for a much better end-user experience if user has to
wait less on each step rather than wait a long time on one step.

3. I did try to change this to only create config before review page
rather than the whole VM. The problem is that we can't create complete
domain configuration before creation of main storage volume as we can
only reliably get the path of the volume *after* the volume is
created. So we'll either have to have a partial domain config at review
and we add the volume's path at creation time (seems pretty ugly to me)
or we create the volume before review page (which brings us to the same
potential issues as with pre-creating the VM).

https://bugzilla.gnome.org/show_bug.cgi?id=674209