installer: Add ::supports_virtio_disk method
This will be useful to be able to automatically use the viostor
driver on Windows when we have it available. We already have
a check for that, but this commit moves it to its own method
that derived classes can override if they want. This will
allow Windows unattended installers to override the info from
libosinfo when the virtio drivers can be provided at install time.

https://bugzilla.gnome.org/show_bug.cgi?id=679752