Avoid deadlock during startup
UnattendedInstaller was doing a sync dbus call in the class
initializer. This is a no-no in GObject because the class initializer
is called with a recursive lock held, and the dbus call will wait
for the dbus thread to reply, but the dbus thread may create a
GObject and thus need the class_init lock, which causes a deadlock.

https://bugzilla.gnome.org/show_bug.cgi?id=676679