Use LayoutManagers and implicit animations
This rewrites the whole layout and animation setup to use Clutter 1.10 style
LayoutManager APIs and implicit animations. It also removes the use of deprecated
APIs like ClutterState, etc.

This leads to less magic numbers and a more robust layout, as well as niceties like
being able to automatically set a minimum size for the window (although this needs
more work).

https://bugzilla.gnome.org/show_bug.cgi?id=677274