Put the dark background inside clutter
We make most main actors have a transparent background, then
we add a clutter actor at the bottom of the stage with the dark background.
This means we see background everywhere, even e.g. under the sidebar, so
that we can e.g. properly scroll it in and have a shadow cast by it.

https://bugzilla.gnome.org/show_bug.cgi?id=677274