libvirt: provide a menu entry to force shutdown
I would prefer to only show the force shutdown entry when the Box
can be shutdown, via libvirt. But it doesn't seem to be easy to
remove or hide an item from the global menu. Also, I wonder if
from a UI POV this is very desirable, since it may or may not be
in the menu.

https://bugzilla.gnome.org/show_bug.cgi?id=677690