Set post install config soon after starting domain
There is no need to wait for domain to shutdown first. Since our VMs are
shared with all apps using libvirt, its a very real scenerio that first
post-install reboot happens while Boxes is not runnnig: User launches the
saved (in-installation) domain from another (libvirt-based) app. This
patch fixes that.

A small bonus of doing this is that it also makes the issue of libvirt
events not getting in[1] much less severe (i-e Fedora installer wont keep
installing over and over again).

[1] https://bugzilla.redhat.com/show_bug.cgi?id=819617

https://bugzilla.gnome.org/show_bug.cgi?id=678888