ovirt: Add a way to set the SPICE host subject
In some circumstances, the certificate the SPICE host we are
connecting to will not be valid for the IP we are connecting to.
When this happens, a 'host subject' has to be provided to spice-gtk
so that it can authenticate the host it's connecting to.
This has been implemented very recently in oVirt 3.1 REST API, see
https://bugzilla.redhat.com/show_bug.cgi?id=807384 but requires
some non-trivial work in libgovirt. In the mean time, add a
BOXES_SPICE_HOST_SUBJECT environment variable for the cases when
it's required, but it's really a hack and should be solved by using
this new REST API.

The host subject looks like O=organization,CN=hostname, where organization
is an arbitrary organization name, and hostname is the hostname/IP
the certificate was generated for. If the host address hasn't changed,
its IP should work. It can happen that the host address isn't the SPICE address
we are connecting to since oVirt can have different management/display addresses.

https://bugzilla.gnome.org/show_bug.cgi?id=681747