Switch to per-actor easing states
Instead of using layout animations on the overlay_bin we
use the new clutter support for per-actor default easing
states. This means we can now disable animations of the
thumbnail in properties view after the initial animation.

Also, its possible that the layout manager animation
feature will be removed, so better not rely on it.

https://bugzilla.gnome.org/show_bug.cgi?id=678545