Fix parsing of CSS stylesheet with gtk+ 3.4
gtk+ 3.4 expects the horizontal and vertical offsets of text-shadow
and icon-shadow CSS properties to be unitless otherwise it fails
to parse them, and it returns an error parsing the whole CSS stylesheet
("Horizontal and vertical offsets are required").
This causes GNOME Boxes to be wrongly styled/themed, and looks ugly.
This will be fixed with gtk+ 3.6, but in the 3.4 series, let's go
with an easy fix and remove units from these properties.

https://bugzilla.gnome.org/show_bug.cgi?id=678876