Fix layout of initial wizard page
The page had no max width, making the label become very
wide (not wrapping) when the window was wide. Also, the
margins were not consisten.

The fix works by setting the minimal size and the margin
on the notebook itself, and properly centering the first
notebook page. We also set a max width on the label.

https://bugzilla.gnome.org/show_bug.cgi?id=678769