vm-creator: More reliable post-installation setup
We need to make sure that post-installation setup is only done once.
App.add_domain() gets called twice (assuming libvirt events are working)
at domain creation time. Then we have the case of this function being
called for existing domains when Boxes starts-up. This function adds the
new machine to collection, that emits 'item_added' signal and ultimately
that trigered the post_installation setup.

We no longer listen to the 'item_added' signal on the collection but
launch the post-installation setup (or hook it to domain shutdown)
where/when its appropriate.

This fixes the recent regression of live VMs failing on first launch. It
also re-adds the launch of post-installation setup for existing domains
on boxes startup.

https://bugzilla.gnome.org/show_bug.cgi?id=678894