Fix nasty flashing when clicking on toolbar
The spice widget is non-double-buffered, which interacts poorly with
transparent windows. What happens in this case is that the EvenBox
that the spice widet is inside is transparent, but its parent (the window
is not). When drawing the window gdk believes the area under the EventBox
is transparent, so its drawn by the window, and then not drawn as transparent
by the EventBox. When we get to the non-double buffered spice widget
we need to flush the current double buffer pixmap which temporarily shows
the window background before drawing the spice contents.

The fix here is to make the EventBox get a solid background. However, I don't
believe this should be necessary. Gdk should know that the spice widgets
window is opaque and fully covers the evenbox, so it should not have
to paint the window background under the eventbox. I believe this should
be fixable in Gtk+.

https://bugzilla.gnome.org/show_bug.cgi?id=679144