ovirt: Implement OvirtMachine::take_screenshot
This makes it possible to reuse the machinery which takes regular
box screenshot. This implementation of take_screenshot is arguably not
really useful since screenshots will only be taken while the box is
maximized/fullscreen, and the disconnect_display () logic takes care of
taking a screenshot just before going back to the collection view. Taking
regular screenshots can have its use in case of an abnormal gnome-boxes
termination.

https://bugzilla.gnome.org/show_bug.cgi?id=681747