Factor out MachineActor.update_display()
In a following patch, we will update the display widget while in
Properties page. That new function take care of that.

Note: I don't know what the Idle was used for, it doesn't seem to
change anything.

Note2: The animation duration seems to be longer now, so I divided it
herer arbitrarily by 2.., I suspect this is related to the multiple
size_alloc() events (and no, the Idle was not helping with this
problem)

https://bugzilla.gnome.org/show_bug.cgi?id=680293