search: add a search bar
The searchbar is based on current gnome-documents behaviour and style,
as well as the same GtkWidget code shared in libgd.

It will search as you type and filter the boxes. You can activate the
result if there is only a single match currently (I am not sure that's
what the design says).

Anyway, it's hopefully a good and useful first iteration.

https://bugzilla.gnome.org/show_bug.cgi?id=681089