display-config: learn to save various access information
We will save a few useful informations for sorting / statistics:
- time spent on a box
- number of time using a box
- last time access
- first time access

I wanted to implement the base class as a mixin, but vala/gobject is a
bit too limited for that sadly. Not a big deal though.

In the future, I would like to extend this approach to support
Zeitgeist informations too. Also, it might be interesting that we save
these information in libvirt XML too, although it's perhpas preferable
to have a Boxes level of thing, that can work with direct or ovirt
connections too.

https://bugzilla.gnome.org/show_bug.cgi?id=681104