wizard: Keep user in 'setup' until all data is provided
We used to fail with an appropriate error message when user tried to hit
'continue' in setup page without providing needed data. Recently we had a
regression in that regard so user was only getting a "VM setup failed"
error.

This patch fixes the regression but with an alternative solution: Keep the
'continue' button disabled as long as user hasn't provided all the needed
data.

Also now that fedora express install doesn't require password, username
is the only field that is mandatory now.

https://bugzilla.gnome.org/show_bug.cgi?id=680825