vm-configuration: Save media & OS info for new VMs
We now put all the custom XML under a toplevel 'gnome-boxes' node. We do
this without breaking compatibility with existing under installation and
live boxes out there.

Currently this info is only there during the installation but the patch
to make it permanent follows in the same patch series as this.

https://bugzilla.gnome.org/show_bug.cgi?id=688770