collection-view: Use GdMainIconView instead of GtkIconView
While rest of the UI looks exactly it did before this change, the
selection view now looks a lot prettier (i-e like in gnome-documents).
Apart from the aesthetic benefits, we now also retain the selection when
switching away from selection view.

Most of this code is taken (and ported to Vala) from GdMainView that we
(at least currently) can not use.

https://bugzilla.gnome.org/show_bug.cgi?id=681911