vm-creator: Track & show installation progress
Its completely based on us knowing the guest disk usage after
installation and we only know that in case of express installation. We
do the same for non-express installations and live VMs.

Another limitation is that we lose track of installations if user quits
boxes in btw installation.

On the bright side, we now have a more reliable way of knowing when
installation is really complete for express installs.

https://bugzilla.gnome.org/show_bug.cgi?id=682300