Download logos on demand
This patch is very similar to the one I originally proposed in
bug#670003. The understanding I got from Fedora legal was that
distribution of the logos is not an issue but rather the use so I started
to ship logos in Boxes itself. While it was really not an issue for
Fedora legal, it became an issue for other distros and upstream
(bug#671251).

I then put this under a new package called gnome-boxes-nonfree but soon
realized that most distros would not like to ship it in their main
repositories as it clearly says 'nonfree' in the name. So I would like
to retire that package and download these logos on demand and keep them
cached.

Major differences against the original patch in bug#670003:

1. Now we ourselves maintain a very tiny libosinfo database adds logo
logo URLs to some OSs. The reason for that is that Fedora legal asked us
to ensure that these logos are not used outside of Boxes' context.

2. I've put all the downloading and caching code in a separate new
singleton class: Downloader. Later I intend to use this to also download
ISOs.

3. Ensure same file is not downloaded more than once.

https://bugzilla.gnome.org/show_bug.cgi?id=682573