Change installation progress string to back to 'Installed'
Commit  f18387885 was meant to use ngettext to translate "%d%% Installed"
but after review, there was an unwanted s/Installed/installed change
that I was supposed to revert. For some reason, I messed this up. This
is just a follow up commit to restore the string to what it should have
been