Ignore almost-black screenshots
Black and almost black screenshots are almost always screensaver
or lock screens. These just make the collection view icons
be full of black screenshots after a while, which is not very
useful. It is also easy to confuse with turned off boxes which
we want to make black.

To fix this we just ignore all screenshot that are too dark.

https://bugzilla.gnome.org/show_bug.cgi?id=683067