Lower the detect-black limit again
The increased value for blackness detection had a lot
of miss-triggering, like in the windows installer, while
not detecting the default Windows screensaver anyway.

So, better safe, we lower it to something that is far
less likely to get false positives. The important thing
is to not allow very dark screens that may be confused
with a turned off box.