Add generic broker support to App
Add some code in App to let brokers register the uri schemes they know
how to handle, and to automatically use the registered brokers. This
is achieved through the addition of a Broker base class.
However, I'm not sure it's possible to have broker registration run
automatically at the application startup, so we may still need a bit
of broker specific code in App.vala :(

https://bugzilla.gnome.org/show_bug.cgi?id=681747