build: add Makefile rule to download extra Win drivers ISO
Since a4a0780c8, Boxes is able to use an ISO with extra Windows
drivers to auto-install the viostor virtio drivers during Windows
unattended installations. However, it does not know yet how to
download that ISO. As a stop-gap measure, add a Makefile rule
to download it and save it with the right name. It can then be
copied to ~/.cache/gnome-boxes/ or to $data_dir/gnome-boxes/, and
Boxes will be able to make use of it.

https://bugzilla.gnome.org/show_bug.cgi?id=679752