Fix zoom in animation when starting box.
When clicking on a box the first time it zooms to the center.
But the second (and after) time it stays in place.

Dropping the deprecated ClutterBinAlignment stuff and using the
new clutter 1.12 ClutterActor.x/y_align properties fixes this.

This also bumps the clutter requirement as this seems to only
work on the most recent versions.

https://bugzilla.gnome.org/show_bug.cgi?id=683424