Fix critical error when exiting display mode early
If you go back to the collection view really fast, before the zoom
from CREDS to DISPLAY is finished you get an critical:

(gnome-boxes:32269): Boxes-CRITICAL **: boxes_display_get_display: assertion `self != NULL' failed

This is because the timeout set up to switch to display mode after the
animation is not cleaned up. Fix this by removing the timeout when
disconnecting the show signal handler.

https://bugzilla.gnome.org/show_bug.cgi?id=683522