Make sure 'QEMU Session' exists at startup
Without it, Boxes fails to start:

(gnome-boxes:19000): Boxes-WARNING **: app.vala:322: No such file or directory
error: missing or failing default libvirt connection
$

At startup, Boxes creates this file if ~/.config/gnome-boxes/sources/
does not exist. It's better to directly check for the file we are
interested in. Reason for this patch is that while testing other patches,
this file went missing...

https://bugzilla.gnome.org/show_bug.cgi?id=683558