Do main channel handlers clean-up
Spice-gtk has a peculiar (and annoying) channel object memory
management model... Channels are actually removed from the session
when the refcount reaches 0. If not, bad interactions with new
channels happen (if the session is recycled on reconnect or migration).

I wish spice-gtk would be fixed to be simpler to use, but Boxes can
also learn to clean up old handlers.

https://bugzilla.gnome.org/show_bug.cgi?id=683561