Report errors in Machine::get_screenshot_filename()
When the VM has just been created, config.uuid is not known yet.
When we quit Boxes from the wizard summary screen, the VM is deleted
when it's in such a state, which causes Criticals on the console when
it tries to delete the VM screenshot:

(gnome-boxes:5027): Boxes-CRITICAL **: boxes_get_screenshot_filename:
assertion `prefix != NULL' failed

This commit changes get_screenshot_filename() to report errors by
throwing a Boxes.Error exception and catches it in callers. This
removes the get_screenshot_filename() criticals in the scenario
described above.

https://bugzilla.gnome.org/show_bug.cgi?id=683792