Fix memory customization
libvirt uses 2 XML elements to set memory, a /domain/memory node,
and a /domain/currentMemory node. The former sets the maximum amount
available, and the latter sets the amount of memory currently
available to the guest. If only /domain/memory is specified, libvirt
will automatically insert a /domain/currentMemory node.
This means that after creating a domain, if we want to change the
amount of RAM available to the guest, both values must be changed,
otherwise /domain/currentMemory will limit the amount of available
memory.
This commit sets the 'current-memory' property in addition to the
'memory' property when changing the memory. This fixes raising the
amount of memory available to the guest. It's done by using GObject
properties instead of directly using
gvir_config_domain_set_current_memory to avoid raising our
libvirt-glib dependency.

https://bugzilla.gnome.org/show_bug.cgi?id=672554