Allow Machine::connect_display to report errors
If an error occurs while we connect to a remote display, we
currently have no way of reporting this to the user. When
this happens, Boxes is stuck on showing a zoomed in box snapshot
in the middle of the screen, and nothing else happens.
This commit allows Machine::connect_display to throw exceptions, and
catches them to report errors. However, this is just preparatory work
as the various Machine::connect_display needs to be modified to actually
report the errors they encounter.

https://bugzilla.gnome.org/show_bug.cgi?id=684224