Reset install_media when it has become stale
When clicking on an ISO from the ISO list, we already know
'install_media' so we can set it immediatly. When we
select a media from the filechooser, we don't know the
corresponding install_media immediatly, so we have to guess
it later on.
However, these 2 different situations conflict when the user
first selects a media from the ISO list in the wizard, then goes
back in the wizard and selects a file from the file chooser.
The old media is then reused instead of the new one being used.
This commit sets install_media to null when a file is selected
in the file chooser so that it's guess later on instead of being
reused.
I chose not to clear install_media when navigating through the
wizard in order to be able to reuse the media when going back in
the wizard, and then forward without changing anything.
An alternate fix that would likely work is to always reset it to
null, and to reset it every time the URI change.

https://bugzilla.gnome.org/show_bug.cgi?id=684219