Add error popup on box connection errors
This commit adds a new string and we are in string freeze break,
so approval is needed to get this change in. However, the patch
series is still a good improvement even without the added string,
so I'm splitting this out to be able to commit the rest of the series
(thanks Alex for the suggestion!).
However, without this error message, Boxes starting to connect to a box
and then giving up can be quite confusing, so adding this message
is a worthwhile user-friendliness improvement.

https://bugzilla.gnome.org/show_bug.cgi?id=684224