Fix multiple connections on startup
This is a better fix for bug 683042 (where we connect twice
when clicking on a box). It fixes the race where we connect twice
as well as fix the regression where password authentication was
broken.

We now handle the race by making connect_display() safe to run
multiple times by:
 * Making update_display() only create a display if needed.
   This is safe as we always clear the display property before
   reconnecting.
 * The Display.connect_it() method was changed to only connect
   once.

Additionally we make reconnect_display() only reconnect if there
is already a display. This means that if we hit reconnect_display() before
the initial connect_display() has started connecting we don't connect.
This ensures connect_display() does the actual connection, so that it
can properly report any errors.