Add line between topbar and icon view
This adds a line between the top bar and the icon view, similar
to what gnome-documents does it. In fact, it uses the same style
class documents-scrolledwin that GdMainView uses for this, which
we don't get due to not using GdMainView.

In the normal case this doesn't really change anything in boxes, however
in selection mode we need this line in order to separate the
toolbar from the icon view as they have the same color there.

https://bugzilla.gnome.org/show_bug.cgi?id=674671