Handle topology node missing from libvirt capabilities
The <topology> node is not always present in libvirt capabilities
XML (happened to me when running virsh capabilities in a F18 VM
running on a F18 host). Boxes does not handle this situation, and
tries to use the NULL capabilities pointer, which causes it to
try to create a box with 0 vcpus, which is rejected by libvirt.

This patch checks for NULL capabilities pointers, and falls back to
not setting vcpu count nor topology when this happens as both attributes
are optional. With this change, I can successfully create a box when Boxes
is running in a VM.

https://bugzilla.gnome.org/show_bug.cgi?id=685377