vm-creator: Set post-install config after install
The main reason we have been setting post-installation configuration at
first domain launch was the scenerio of first post-install reboot
happening while Boxes is not runnnig: User launches the saved
(in-installation) domain from another (libvirt-based) app.

While this scenerio is still real, it is:

a. a corner case.
b. very hard to maintain *proper* support for if we want to be able to
   better track the installation. The next patch in this series is along
   that line and it requires this change.
c. not complete broken by this patch: Boxes will just not know that
   installation is finished until another reboot/shutdown.

https://bugzilla.gnome.org/show_bug.cgi?id=685346