wizard: Separate buttons for 'Create' & 'Continue'
Next button changes its size when its label is changed from/to "Continue"
to/from "Create", which could be a big problem depending on the locale
of the user. Imagine user clicking very rapidly on the button to quickly
create a box and hitting 'Back' button at the last page even though
she/he never moved the cursor.

This patche attempts to fix the problem by keeping separate buttons for
both 'Continue' and 'Create' and instead of changing the labels, it
simply shows and hides the buttons appropriately. Since both buttons are
in the same size group, the size of buttons remains the same.

https://bugzilla.gnome.org/show_bug.cgi?id=685779