win7,express: Set UILanguage according to user's locale
We have been hardcoding 'UILanguage' value to make express installation
work for the case of user with a non-US locale using US install media.
This however is a less likely scenerio (and therefore lower priority)
than the case of user with non-US locale using install media of the same
non-US locale. This patch fixes express installation for the latter case
at the cost of breaking it for the former.

Thanks to Christophe Fergeau for testing and finding this out.

https://bugzilla.gnome.org/show_bug.cgi?id=685826