Add storage pool test to --checks
When trying to move the Boxes libvirt storage pool to another place,
Boxes will get non-functional if after the move libvirt configuration
points to an invalid path for the storage pool.
This commit gets the path to Boxes storage pool through 'virsh
pool-dumpxml' and then makes sure the storage pool path is a
user-writable directory.

https://bugzilla.gnome.org/show_bug.cgi?id=676340