Attempt to start Boxes pool if it's inactive
At startup, we try to refresh Boxes storage pool, but this will
fail if it's inactive. This can happen if the user fiddled with
libvirt storage pool outside of Boxes. It's easy to start it so
let's do it if needed.

https://bugzilla.gnome.org/show_bug.cgi?id=676340
(cherry picked from commit 759bda1520bc30074d005d73a311ee771c2d7c51)