search: Fix use-after-free in dbus methods
The async dbus method implementation gets array arguments which
are used after yield, but the generated dbus code that calls
these frees the arguments directly after the call, so we access
them after they are freed.

The correct way to solve this is to make the parameters "owned"
which means the caller needs to copy the arguments. However, this
doesn't seem to work for the case of the dbus generated call, so
we wrap the dbus method implementations in a normal method which
causes owned arguments to work.
(cherry picked from commit 30c63058f6fdd12a0db795b896df553ceeb7891d)