collection: Enter selection mode on right-click
Lets handle right click the same way as Documents does: If there is a
box under the cursor, enter selection mode and select the box.

Note: Unlike Documents, we are not resetting selection when exiting
selection mode and this patch makes this bug even more annoying so that
should be the next thing to fix. Bug#686833 is about that.

https://bugzilla.gnome.org/show_bug.cgi?id=674674