Machine: Always disconnect display when we're not displaying it
We used to only disconnect Machine.display when going into collection
view, but that is not always what happens. For instance, we might
go directly to the wizard if we use the "New" app menu item.
So, we move this disconnect from App.ui_state_changed to
Machine.ui_state_changed and disconnect on any state except
DISPLAY, PROPERTIES and CREDS.

https://bugzilla.gnome.org/show_bug.cgi?id=677714