Break out machine connection into App.connect_to
This merges all the code that handles connection and going into
the CREDS state into a single function instead of being distributed
into App.select_item() and CollectionView UIState.CREDS handling.
This allows connecting to a machine without the collection view being
involved (such as when starting a box from the properties page).

It also moves the place where we collect the actor animation start
position to App.select_item() so that we can use a different start
in other cases.

https://bugzilla.gnome.org/show_bug.cgi?id=677714