Don't crash when running Boxes while it's already running
When running Boxes while an instance is already running, the
new instance crashes in App::suspend_machines. This happens because
App::collection is null and App::suspend_machines tries to dereference
it. This commit just silently exits from App::suspend_machines in such
situations as there will be no machines to save anyway.

https://bugzilla.gnome.org/show_bug.cgi?id=688789