Handle stale file references from tracker
Tracker may give us old data, like if a file was moved/deleted and
not yet rescanned. Currently this makes us give some hard to understand
error when selecting the iso where the file doesn't exist. We now
ensure the file is availible, so such files will not be listed in
the new box wizard.